Work on generalized data sources

+++ src/main/java/org/hdf5javalib/HdfFileReader.java
@@ -43,22 +43,22 @@ public class HdfFileReader {
 //
         // Get the object header address from the superblock
         // Parse the object header from the file using the superblock information
-        long objectHeaderAddress = superblock.getRootGroupSymbolTableEntry().getObjectHeaderAddress().getInstance(BigInteger.class).longValue();
+        long objectHeaderAddress = superblock.getRootGroupSymbolTableEntry().getObjectHeaderAddress().getInstance(Long.class);
         fileChannel.position(objectHeaderAddress);
         HdfObjectHeaderPrefixV1 objectHeader = HdfObjectHeaderPrefixV1.readFromFileChannel(fileChannel, offsetSize, lengthSize);
 
         // Parse the local heap using the file channel
         // Read data from file channel starting at the specified position
-        long localHeapAddress = superblock.getRootGroupSymbolTableEntry().getNameHeapAddress().getInstance(BigInteger.class).longValue();
+        long localHeapAddress = superblock.getRootGroupSymbolTableEntry().getNameHeapAddress().getInstance(Long.class);
         fileChannel.position(localHeapAddress);
         HdfLocalHeap localHeap = HdfLocalHeap.readFromFileChannel(fileChannel, superblock.getOffsetSize(), superblock.getLengthSize());
 
-        long dataSize = localHeap.getDataSegmentSize().getInstance(BigInteger.class).longValue();
-        long dataSegmentAddress = localHeap.getDataSegmentAddress().getInstance(BigInteger.class).longValue();
+        long dataSize = localHeap.getDataSegmentSize().getInstance(Long.class);
+        long dataSegmentAddress = localHeap.getDataSegmentAddress().getInstance(Long.class);
         fileChannel.position(dataSegmentAddress);
         HdfLocalHeapContents localHeapContents = HdfLocalHeapContents.readFromFileChannel(fileChannel, (int) dataSize);
 
-        long bTreeAddress = superblock.getRootGroupSymbolTableEntry().getBTreeAddress().getInstance(BigInteger.class).longValue();
+        long bTreeAddress = superblock.getRootGroupSymbolTableEntry().getBTreeAddress().getInstance(Long.class);
         fileChannel.position(bTreeAddress);
         HdfBTreeV1 bTree = HdfBTreeV1.readFromFileChannel(fileChannel, superblock.getOffsetSize(), superblock.getLengthSize());
 
@@ -67,7 +67,7 @@ public class HdfFileReader {
             throw new UnsupportedEncodingException("Only one btree entry is currently supported");
         }
         HdfBTreeEntry hdfBTreeEntry = bTree.getEntries().get(0);
-        long snodAddress = hdfBTreeEntry.getChildPointer().getInstance(BigInteger.class).longValue();
+        long snodAddress = hdfBTreeEntry.getChildPointer().getInstance(Long.class);
         fileChannel.position(snodAddress);
         HdfGroupSymbolTableNode symbolTableNode = HdfGroupSymbolTableNode.readFromFileChannel(fileChannel, offsetSize);
         int entriesToRead = symbolTableNode.getNumberOfSymbols();
@@ -95,7 +95,7 @@ public class HdfFileReader {
             HdfString datasetName = localHeapContents.parseStringAtOffset(ste.getLinkNameOffset());
             // Parse the Data Object Header Prefix next in line
 //        fileChannel.position(fileOffsets.getObjectHeaderAddress().toBigInteger().longValue());
-            long dataLObjectHeaderAddress = ste.getObjectHeaderAddress().getInstance(BigInteger.class).longValue();
+            long dataLObjectHeaderAddress = ste.getObjectHeaderAddress().getInstance(Long.class);
             fileChannel.position(dataLObjectHeaderAddress);
             dataObjectHeaderPrefix = HdfObjectHeaderPrefixV1.readFromFileChannel(fileChannel, offsetSize, lengthSize);
             System.out.println(datasetName + "@" + dataLObjectHeaderAddress + " = " + dataObjectHeaderPrefix);
@@ -104,10 +104,10 @@ public class HdfFileReader {
                 if (message instanceof DatatypeMessage dataTypeMessage) {
                     dataType = dataTypeMessage.getHdfDatatype();
                 } else if (message instanceof DataLayoutMessage dataLayoutMessage) {
-                    dataAddress = dataLayoutMessage.getDataAddress().getInstance(BigInteger.class).longValue();
-                    dimensionSize = dataLayoutMessage.getDimensionSizes()[0].getInstance(BigInteger.class).longValue();
+                    dataAddress = dataLayoutMessage.getDataAddress().getInstance(Long.class);
+                    dimensionSize = dataLayoutMessage.getDimensionSizes()[0].getInstance(Long.class);
                 } else if (message instanceof DataspaceMessage dataSpaceMessage) {
-                    dimension = dataSpaceMessage.getDimensions()[0].getInstance(BigInteger.class).longValue();
+                    dimension = dataSpaceMessage.getDimensions()[0].getInstance(Long.class);
                 }
             }
